requirements.txt
 
Flask==2.2.2
SQLAlchemy==1.4.41
mysql-connector-python==8.0.33
coverage